Started working a word similarity model
Started a word association model that we can use for our answers. I am not sure if it is better to train 1 model on every story or a smaller model on each story. I am assuming 1 model on every story since we get more data but I will need to do some test to see what works better